IMP: Refactor code.
